linking .css
Finally managed to mirror index.css after index.scss as well as making al the shortcuts from partials to the index.scss.
Also change link to index.css as well as logo on the contact and project page since this was a copy from index.html which is in the root folder, i.e. another linkage to the other files.